E3_model_AI/requirements.txt
pandas
numpy
scikit-learn
xgboost
openpyxl
python-dotenv
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytest
httpx
pytest-cov
pytest-asyncio
opencv-python
Pillow
ultralytics
playwright
streamlit
requests
joblib
beautifulsoup4